+++ 666foodDelivery/Properties/launchSettings.json
@@ -13,7 +13,7 @@
       "launchBrowser": true,
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development",
-        "AZURE_USERNAME": "TP046721@mail.apu.edu.my"
+        "AZURE_USERNAME": "TP046641@mail.apu.edu.my"
       }
     },
     "_666foodDelivery": {
@@ -22,7 +22,7 @@
       "applicationUrl": "https://localhost:5001;http://localhost:5000",
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development",
-        "AZURE_USERNAME": "TP046721@mail.apu.edu.my"
+        "AZURE_USERNAME": "TP046641@mail.apu.edu.my"
       }
     }
   }
