+++ 666foodDelivery/Views/Foods/Create.cshtml
@@ -11,10 +11,10 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form method="post" asp-controller="Blob" asp-action="UploadBlob">
+        <form method="post" enctype="multipart/form-data" asp-controller="Foods" asp-action="UploadBlob">
             <div class="form-group">
                 <label>Upload Image</label>
-                <input class="form-control" name="files" type="image" />
+                <input class="form-control" name="files" type="file" />
             </div>
             <div class="form-group">
                 <input class="btn btn-success" type="submit" value="Submit" />
@@ -24,32 +24,32 @@
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group">
                 <label asp-for="FoodName" class="control-label"></label>
-                <input asp-for="FoodName" class="form-control" />
+                <input asp-for="FoodName" class="form-control" required/>
                 <span asp-validation-for="FoodName" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <label asp-for="FoodProducedDate" class="control-label"></label>
-                <input asp-for="FoodProducedDate" class="form-control" />
+                <input asp-for="FoodProducedDate" class="form-control" required/>
                 <span asp-validation-for="FoodProducedDate" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <label asp-for="Type" class="control-label"></label>
-                <input asp-for="Type" class="form-control" />
+                <input asp-for="Type" class="form-control" required/>
                 <span asp-validation-for="Type" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <label asp-for="Price" class="control-label"></label>
-                <input asp-for="Price" class="form-control" />
+                <input asp-for="Price" class="form-control" required/>
                 <span asp-validation-for="Price" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <label asp-for="Ingredient" class="control-label"></label>
-                <input asp-for="Ingredient" class="form-control" />
+                <input asp-for="Ingredient" class="form-control" required/>
                 <span asp-validation-for="Ingredient" class="text-danger"></span>
             </div>
             <div class="form-group">
                 <label asp-for="BlobURL" class="control-label"></label>
-                <input asp-for="BlobURL" class="form-control" values="@ViewData["BlobUrl"]" />
+                <input asp-for="BlobURL" class="form-control" value="@ViewData["BlobUrl"]" required/>
                 <span asp-validation-for="BlobURL" class="text-danger"></span>
             </div>
             <div class="form-group">
