+++ 666foodDelivery/Views/Products/Details.cshtml
@@ -1,0 +1,9 @@
+﻿@model _666foodDelivery.Models.Food
+
+@{
+    ViewData["Title"] = "Details";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+<h1>Details</h1>
+
