+++ 666foodDelivery/Views/Shared/_LoginPartial.cshtml
@@ -16,7 +16,7 @@
             </form>
         </li>
         <li class="nav-item">
-            <a class="nav-link text-dark" asp-area="" asp-controller="Foods" asp-action="Index">Food Products</a>
+            <a class="nav-link text-dark" asp-area="" asp-controller="Products" asp-action="Index">Food Products</a>
         </li>
     }
     else
